[optimize]調整MainMenu樣式
